FIX: Resolved issue with Jenkins pipeline not having property files for the build
